Extracted common test functionality to avoid code duplication.
